[MCOBERTURA-86] no coverage reported for integration-test Part of this commit was submitted by: Stevo Slavic
Add two new mojos that extend previous mojos, but that change the execution phase from "test" to "verify". This way the integration tests are run as well as the unit tests, making it possible to get coverage for integration tests.

There is no separation of results or reports in these new mojos, so they will only work for people who want aggregated test coverage numbers for both unit tests and integration tests.

git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/cobertura-maven-plugin@20181 52ab4f32-60fc-0310-b215-8acea882cd1b