[MCOBERTURA-86] no coverage reported for integration-test Use only the old cobertura report in the aggregate report integration test, to hopefully fix the failure in Bamboo. The failure only exists when run on Maven 2. With Maven 3 the test passes.
The reason for the failure is that the first instrumentation alters the outputDirectory pointing it to the instrumented classes so that later plugins will pick those up. My understanding is that this changed outputDirectory persists, under Maven 2, when the second instrumentation starts (for integration tests). The failure occurs when the plugin tries to copy the compiled classes to the instrumentation directory, but under Maven 2 this means that the source and target directory are the same, both pointing to the instrumentation directory.

git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/cobertura-maven-plugin@20192 52ab4f32-60fc-0310-b215-8acea882cd1b